+++ application/views/contract/index.php
@@ -22,7 +22,9 @@
                 Contract
                 <div class="pull-right">
                     <div class="btn-group">
+                        <?php if($this->session->userdata('cont_add') == 1) {?> 
                         <a class="btn btn-default btn-xs" href="<?=base_url()?>contract/add"><i class="fa fa-plus"></i> Add</a>
+                        <?php } ?>
                     </div>
                 </div>
             </div>
@@ -64,8 +66,13 @@
                                     <td><?=$contract->contract_staff?></td>
                                     <th><?=$contract->contract_remarks?></th>
                                     <td>
-                                        <a title="Edit" href="<?=base_url()?>contract/edit/<?=$contract->contract_id?>"><i class="fa fa-pencil-square-o"></i></a>&nbsp
-                                        <a title="Delete" href="<?=base_url()?>contract/delete/<?=$contract->contract_id?>" onclick="return confirm_delete()" ><i class="fa fa-trash-o"></i></a>
+                                          <?php if($this->session->userdata('cont_edit') == 1) {?>
+                                         <a title="Edit" href="<?=base_url()?>contract/edit/<?=$contract->contract_id?>"><i class="fa fa-pencil-square-o"></i></a>&nbsp
+                                         <?php } ?>
+
+                                          <?php if($this->session->userdata('cont_del') == 1) {?> 
+                                             <a title="Delete" href="<?=base_url()?>contract/delete/<?=$contract->contract_id?>" onclick="return confirm_delete()" ><i class="fa fa-trash-o"></i></a>
+                                         <?php } ?>
                                     </td>
                                 </tr>
                                 <?php endforeach ?>
